PAYEE-729: Rename classes so they correspond to the filename. Split EstimatedIncomeViewModel into multiple classes.
