+++ app/views/benefits/RemoveCompanyBenefitCheckYourAnswersView.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import uk.gov.hmrc.tai.viewModels.benefit.RemoveCompanyBenefitCheckYourAnswersViewModel
+@import uk.gov.hmrc.tai.viewModels.benefit.RemoveCompanyBenefitsCheckYourAnswersViewModel
 @import includes.ptaHeader
 @import uk.gov.hmrc.play.views.html.helpers.FormWithCSRF
 @import scala.concurrent.ExecutionContext
@@ -24,7 +24,7 @@
 
 @this(main: MainTempleteNew, form: FormWithCSRF)
 
-@(viewModel: RemoveCompanyBenefitCheckYourAnswersViewModel)(implicit request: Request[_], messages: Messages, user: controllers.auth.AuthedUser, templateRenderer: uk.gov.hmrc.renderer.TemplateRenderer, ec: ExecutionContext)
+@(viewModel: RemoveCompanyBenefitsCheckYourAnswersViewModel)(implicit request: Request[_], messages: Messages, user: controllers.auth.AuthedUser, templateRenderer: uk.gov.hmrc.renderer.TemplateRenderer, ec: ExecutionContext)
 
 @header = @{
   ptaHeader(
