+++ test/controllers/TaxAccountSummaryControllerSpec.scala
@@ -23,11 +23,11 @@ import org.mockito.Matchers.any
 import org.mockito.Mockito.{times, verify, when}
 import org.mockito.{Matchers, Mockito}
 import org.scalatest.BeforeAndAfterEach
-import play.api.test.Helpers.{contentAsString, status, _}
+import play.api.test.Helpers._
 import uk.gov.hmrc.http.UnauthorizedException
 import uk.gov.hmrc.play.audit.http.connector.AuditResult.Success
 import uk.gov.hmrc.tai.connectors.responses.{TaiNotFoundResponse, TaiSuccessResponseWithPayload, TaiTaxAccountFailureResponse, TaiUnauthorisedResponse}
-import uk.gov.hmrc.tai.model.IncomesSources
+import uk.gov.hmrc.tai.model.IncomeSources
 import uk.gov.hmrc.tai.model.domain._
 import uk.gov.hmrc.tai.model.domain.income._
 import uk.gov.hmrc.tai.service._
@@ -58,7 +58,7 @@ class TaxAccountSummaryControllerSpec
             taxAccountSummary,
             ThreeWeeks,
             nonTaxCodeIncome,
-            IncomesSources(livePensionIncomeSources, liveEmploymentIncomeSources, ceasedEmploymentIncomeSources),
+            IncomeSources(livePensionIncomeSources, liveEmploymentIncomeSources, ceasedEmploymentIncomeSources),
             nonMatchedEmployments
           ))
       )
@@ -84,7 +84,7 @@ class TaxAccountSummaryControllerSpec
             taxAccountSummary,
             ThreeWeeks,
             nonTaxCodeIncome,
-            IncomesSources(livePensionIncomeSources, liveEmploymentIncomeSources, ceasedEmploymentIncomeSources),
+            IncomeSources(livePensionIncomeSources, liveEmploymentIncomeSources, ceasedEmploymentIncomeSources),
             nonMatchedEmployments
           ))
       )
@@ -111,7 +111,7 @@ class TaxAccountSummaryControllerSpec
               taxAccountSummary,
               ThreeWeeks,
               nonTaxCodeIncome,
-              IncomesSources(livePensionIncomeSources, liveEmploymentIncomeSources, ceasedEmploymentIncomeSources),
+              IncomeSources(livePensionIncomeSources, liveEmploymentIncomeSources, ceasedEmploymentIncomeSources),
               nonMatchedEmployments
             ))
         )
